refactor: #58 remove unused files, organise folder structure for characterdetail
