+++ resources/js/Components/CharacterDetail/CharacterDetailBox.tsx
