+++ resources/js/Components/CharacterDetail/index.ts
@@ -1,0 +1,3 @@
+export {default as CharacterDetailBox} from './CharacterDetailBox'
+export {default as CharacterDetailTable} from './CharacterDetailTable'
+export {default as EnhancedTableToolbar} from './EnhancedTableToolbar'
