--- resources/js/Components/NavLink.tsx
@@ -1,24 +1,0 @@
-/* eslint-disable */
-// These files will be deleted once login is implemented
-import {InertiaLink} from '@inertiajs/inertia-react'
-import React from 'react'
-
-interface Props {
-    href: string
-    active: boolean
-    children: React.ReactNode
-}
-
-export default function NavLink({href, active, children}: Props) {
-    return (
-        <InertiaLink
-            href={href}
-            className={
-                active
-                    ? 'inline-flex items-center px-1 pt-1 border-b-2 border-indigo-400 text-sm font-medium leading-5 text-gray-900 focus:outline-none focus:border-indigo-700 transition duration-150 ease-in-out'
-                    : 'inline-flex items-center px-1 pt-1 border-b-2 border-transparent text-sm font-medium leading-5 text-gray-500 hover:text-gray-700 hover:border-gray-300 focus:outline-none focus:text-gray-700 focus:border-gray-300 transition duration-150 ease-in-out'
-            }>
-            {children}
-        </InertiaLink>
-    )
-}
