+++ resources/js/Components/index.ts
@@ -1,2 +1,8 @@
 export {default as Authentication} from './Authentication'
 export {default as FactionChip} from './FactionChip'
+
+export {
+    CharacterDetailBox,
+    CharacterDetailTable,
+    EnhancedTableToolbar,
+} from './CharacterDetail'
