+++ resources/js/Pages/Character/Detail/CharacterDetail.test.js
