+++ resources/js/Pages/Character/Detail/index.ts
