--- resources/js/Pages/Dashboard.tsx
@@ -1,26 +1,0 @@
-/* eslint-disable */
-// These files will be deleted once login is implemented
-import React from 'react'
-import Authenticated from '@/Layouts/Authenticated'
-
-export default function Dashboard(props: React.PropsWithChildren<any>) {
-    return (
-        <Authenticated
-            auth={props.auth}
-            header={
-                <h2 className='font-semibold text-xl text-gray-800 leading-tight'>
-                    Dashboard
-                </h2>
-            }>
-            <div className='py-12'>
-                <div className='max-w-7xl mx-auto sm:px-6 lg:px-8'>
-                    <div className='bg-white overflow-hidden shadow-sm sm:rounded-lg'>
-                        <div className='p-6 bg-white border-b border-gray-200'>
-                            You're logged in!
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </Authenticated>
-    )
-}
