+++ resources/mock/character-detail-data.tsx
